@@ -8,9 +8,9 @@
   "license": "MIT",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack --config webpack.prod.js",
+    "build": "webpack --config webpack.prod.mjs",
     "lint": "eslint ./src/**/* --ext .js,.jsx,.ts,.tsx",
-    "start": "webpack-dev-server --open --config webpack.dev.js"
+    "start": "webpack serve --open --config webpack.dev.mjs"
   },
   "files": [
     "src/**/*",
@@ -23,31 +23,28 @@
   ],
   "style": "src/style.css",
   "dependencies": {
-    "@types/cash": "^0.0.3",
-    "cash-dom": "^7.0.1",
+    "cash-dom": "^8.1.5",
     "esprima": "^4.0.1"
   },
   "devDependencies": {
-    "@types/esprima": "^4.0.2",
-    "@typescript-eslint/eslint-plugin": "^2.31.0",
-    "@typescript-eslint/parser": "^2.31.0",
-    "autoprefixer": "^9.7.6",
-    "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^5.1.1",
-    "css-loader": "^3.5.3",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb-typescript": "^7.2.1",
-    "eslint-plugin-import": "^2.20.2",
-    "html-webpack-plugin": "^4.3.0",
-    "less": "^3.11.1",
-    "less-loader": "^6.0.0",
-    "mini-css-extract-plugin": "^0.9.0",
-    "postcss-loader": "^3.0.0",
-    "ts-loader": "^7.0.2",
-    "typescript": "^3.8.3",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3",
-    "webpack-merge": "^4.2.2"
+    "@types/codemirror": "^5.60.2",
+    "@types/esprima": "^4.0.6",
+    "autoprefixer": "^10.4.21",
+    "clean-webpack-plugin": "^4.0.0",
+    "copy-webpack-plugin": "^13.0.1",
+    "css-loader": "^7.1.2",
+    "eslint": "^9.35.0",
+    "html-webpack-plugin": "^5.6.4",
+    "less": "^4.4.1",
+    "less-loader": "^12.3.0",
+    "mini-css-extract-plugin": "^2.9.4",
+    "postcss-loader": "^8.2.0",
+    "ts-loader": "^9.5.4",
+    "typescript": "^5.9.2",
+    "typescript-eslint": "^8.44.0",
+    "webpack": "^5.101.3",
+    "webpack-cli": "^6.0.1",
+    "webpack-dev-server": "^5.2.2",
+    "webpack-merge": "^6.0.1"
   }
 }
